tonic_reflection::pb
Module v1

Generated protobuf types from the grpc.reflection.v1 package.

Modules

server_reflection_client

Generated client implementations.

server_reflection_request

Nested message and enum types in ServerReflectionRequest.

server_reflection_response

Nested message and enum types in ServerReflectionResponse.

server_reflection_server

Generated server implementations.

Structs

ErrorResponse

The error code and error message sent by the server when an error occurs.

ExtensionNumberResponse

A list of extension numbers sent by the server answering all_extension_numbers_of_type request.

ExtensionRequest

The type name and extension number sent by the client when requesting file_containing_extension.

FileDescriptorResponse

Serialized FileDescriptorProto messages sent by the server answering a file_by_filename, file_containing_symbol, or file_containing_extension request.

ListServiceResponse

A list of ServiceResponse sent by the server answering list_services request.

ServerReflectionRequest

The message sent by the client when calling ServerReflectionInfo method.

ServerReflectionResponse

The message sent by the server to answer ServerReflectionInfo method.

ServiceResponse

The information of a single service used by ListServiceResponse to answer list_services request.

Constants

FILE_DESCRIPTOR_SET

Byte encoded FILE_DESCRIPTOR_SET.