Struct tonic_reflection::pb::v1::ServerReflectionResponse

pub struct ServerReflectionResponse {
    pub valid_host: String,
    pub original_request: Option<ServerReflectionRequest>,
    pub message_response: Option<MessageResponse>,
}

The message sent by the server to answer ServerReflectionInfo method.

Fields
valid_host: Stringoriginal_request: Option<ServerReflectionRequest>message_response: Option<MessageResponse>

The server sets one of the following fields according to the message_request in the request.