tonic_reflection::pb::v1alpha

Module server_reflection_request

Nested message and enum types in ServerReflectionRequest.

Enums

MessageRequest

To use reflection service, the client should set one of the following fields in message_request. The server distinguishes requests by their defined field and then handles them using corresponding methods.