tonic_reflection::pb::v1alpha

Struct ServiceResponse

pub struct ServiceResponse {
    pub name: String,
}

The information of a single service used by ListServiceResponse to answer list_services request.

Fields
name: String

Full name of a registered service, including its package name. The format is .