tonic_reflection::server::v1alpha

Trait ServerReflection

pub trait ServerReflection:
    Send
    + Sync
    + 'static {
    type ServerReflectionInfoStream: Stream<Item = Result<ServerReflectionResponse, Status>> + Send + 'static;

    // Required method
    fn server_reflection_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Streaming<ServerReflectionRequest>>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ServerReflectionInfoStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ServerReflectionServer.

Required Associated Types

type ServerReflectionInfoStream: Stream<Item = Result<ServerReflectionResponse, Status>> + Send + 'static

Server streaming response type for the ServerReflectionInfo method.

Required Methods

fn server_reflection_info<'life0, 'async_trait>( &'life0 self, request: Request<Streaming<ServerReflectionRequest>>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::ServerReflectionInfoStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

The reflection service is structured as a bidirectional stream, ensuring all related requests go to a single server.